Fix pgsql build failures (#783)
The pgsql test suite (recording mode against postgres:14) failed on 4
assertions while every mysql job stayed green. PostgreSQL's strict typing
rejects simulated SQL that MySQL silently casts, exposing three issues:

- Positional '?' parameters were substituted in the wrong order. Bind
  calls are discovered in reverse source order by ExpressionFinder, so
  `WHERE a = ? and b = ?` became `a = '<b-value>' and b = '<a-value>'`.
  MySQL tolerated the mismatched literal, pgsql could not cast it. Sort
  the parameters by placeholder key before substitution.

- The `number` type (int|float) was simulated as '1.0', which pgsql
  refuses to cast to an integer column. Prefer an integer representative
  for unions, which is accepted by both integer and float/decimal columns.

- The CTE tests added in #776 were mysql-only: `gesperrt int<-128,127>`
  assumes mysql TINYINT (pgsql schema uses INT) and `CAST(... AS SIGNED)`
  is invalid pgsql syntax. Keep the cross-db cases shared and split the
  aggregation/recursive cases into engine-specific data files.
